typescript(vx-demo): rewrite BarGroupHorizontal example in TypeScript
